refactor(dbt): rename `cli/` to `core/`
