+++ python_modules/libraries/dagster-dbt/dagster_dbt_tests/test_dependencies.py
@@ -7,7 +7,7 @@ from dagster import (
 )
 from dagster._core.errors import DagsterInvalidInvocationError
 from dagster_dbt.asset_decorator import dbt_assets
-from dagster_dbt.cli import DbtManifest
+from dagster_dbt.core.resources_v2 import DbtManifest
 
 test_dagster_metadata_manifest = DbtManifest.read(
     path=Path(__file__).parent.joinpath("dbt_projects", "test_dagster_metadata", "manifest.json")
